% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_plot.R
\name{fp_plot}
\alias{fp_plot}
\title{Plot logarithmic returns of a stock series}
\usage{
fp_plot(data)
}
\arguments{
\item{data}{A long dataframe containing dates and adjusted closing prices for stocks, first column is date and second column is the ticker and 3rd column is adjusted close price.}
}
\value{
ggplot object that illustrates historical performance of a stocks returns.
}
\description{
Plot logarithmic returns of a stock series
}
\examples{
fp_plot(stocks)

}
