% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_regress.R
\name{roll_regress}
\alias{roll_regress}
\title{Perform Linear Regression with Rolling Window Size}
\usage{
roll_regress(data, window)
}
\arguments{
\item{data}{A Wide data frame with n columns containing log returns, first column is date, second is the independent variable returns, and the rest are dependent variables' returns.}

\item{window}{The window size in days for which to run a regression on. 30 means it will run a regression on days 1-30 then 2-31 and so on}
}
\value{
A table with summary statistics for each regression pair
}
\description{
Perform Linear Regression with Rolling Window Size
}
\examples{

\dontrun{roll_regress(stocks_2, window = 60)}
}
