% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stocks}
\alias{stocks}
\title{Stock prices dataset}
\format{
A tibble with 3 variables:
\describe{
  \item{date}{Trading date}
  \item{series}{Stock ticker symbol}
  \item{value}{Adjusted closing price}
}
}
\source{
Yahoo Finance via tidyquant
}
\usage{
stocks
}
\description{
Adjusted stock prices for SPY
}
\keyword{datasets}
