% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stocks_2}
\alias{stocks_2}
\title{Stock prices dataset for 2 stocks}
\format{
A tibble with 3 variables:
\describe{
  \item{date}{Trading date}
  \item{series}{Stock ticker symbol}
  \item{value}{Adjusted closing price}
}
}
\source{
Yahoo Finance via tidyquant
}
\usage{
stocks_2
}
\description{
Adjusted stock prices for SPY and AAPL
}
\keyword{datasets}
